+++ src/lightning_habana/pytorch/strategies/parallel.py
@@ -17,38 +17,55 @@ from typing import Any, Callable, Dict, List, Optional, Union
 import torch
 import torch.distributed
 from lightning_utilities import module_available
-
-if module_available("lightning"):
-    from lightning.fabric.plugins import CheckpointIO, ClusterEnvironment
-    from lightning.fabric.utilities.distributed import group as _group
-    from lightning.fabric.utilities.types import ReduceOp
-    from lightning.pytorch import LightningModule
-    from lightning.pytorch.accelerators import Accelerator
-    from lightning.pytorch.plugins.io.wrapper import _WrappingCheckpointIO
-    from lightning.pytorch.plugins.precision import PrecisionPlugin
-    from lightning.pytorch.strategies.ddp import DDPStrategy
-    from lightning.pytorch.utilities.rank_zero import rank_zero_warn
-    from lightning.pytorch.utilities.types import STEP_OUTPUT
-elif module_available("pytorch_lightning"):
-    from lightning_fabric.plugins import CheckpointIO, ClusterEnvironment
-    from lightning_fabric.utilities.distributed import group as _group
-    from lightning_fabric.utilities.types import ReduceOp
-    from pytorch_lightning import LightningModule
-    from pytorch_lightning.accelerators import Accelerator
-    from pytorch_lightning.plugins.io.wrapper import _WrappingCheckpointIO
-    from pytorch_lightning.plugins.precision import PrecisionPlugin
-    from pytorch_lightning.strategies.ddp import DDPStrategy
-    from pytorch_lightning.utilities.rank_zero import rank_zero_warn
-    from pytorch_lightning.utilities.types import STEP_OUTPUT
-else:
-    raise ModuleNotFoundError("You are missing `lightning` or `pytorch-lightning` package, please install it.")
+from lightning_utilities.core.rank_zero import rank_zero_only as utils_rank_zero_only
 from torch import Tensor
 from torch.nn import Module
 from torch.optim.optimizer import Optimizer
 
+if module_available("lightning"):
+    from lightning.fabric.plugins import CheckpointIO, ClusterEnvironment
+    from lightning.fabric.plugins.collectives.torch_collective import default_pg_timeout
+    from lightning.fabric.utilities.distributed import (
+        _distributed_is_initialized,
+        _init_dist_connection,
+    )
+    from lightning.fabric.utilities.distributed import group as _group
+    from lightning.fabric.utilities.seed import reset_seed
+    from lightning.fabric.utilities.types import ReduceOp
+    from lightning.pytorch import LightningModule
+    from lightning.pytorch.accelerators import Accelerator
+    from lightning.pytorch.plugins.io.wrapper import _WrappingCheckpointIO
+    from lightning.pytorch.plugins.precision import Precision
+    from lightning.pytorch.strategies.launchers import _MultiProcessingLauncher, _SubprocessScriptLauncher
+    from lightning.pytorch.strategies.parallel import ParallelStrategy
+    from lightning.pytorch.strategies.strategy import TBroadcast
+    from lightning.pytorch.utilities.rank_zero import rank_zero_only
+    from lightning.pytorch.utilities.types import STEP_OUTPUT
+elif module_available("pytorch_lightning"):
+    from lightning_fabric.plugins import CheckpointIO, ClusterEnvironment
+    from lightning_fabric.plugins.collectives.torch_collective import default_pg_timeout
+    from lightning_fabric.utilities.distributed import (
+        _distributed_is_initialized,
+        _init_dist_connection,
+    )
+    from lightning_fabric.utilities.distributed import group as _group
+    from lightning_fabric.utilities.seed import reset_seed
+    from lightning_fabric.utilities.types import ReduceOp
+    from pytorch_lightning import LightningModule
+    from pytorch_lightning.accelerators import Accelerator
+    from pytorch_lightning.plugins.io.wrapper import _WrappingCheckpointIO
+    from pytorch_lightning.plugins.precision import Precision
+    from pytorch_lightning.strategies.launchers import _MultiProcessingLauncher, _SubprocessScriptLauncher
+    from pytorch_lightning.strategies.parallel import ParallelStrategy
+    from pytorch_lightning.strategies.strategy import TBroadcast
+    from pytorch_lightning.utilities.rank_zero import rank_zero_only
+    from pytorch_lightning.utilities.types import STEP_OUTPUT
+else:
+    raise ModuleNotFoundError("You are missing `lightning` or `pytorch-lightning` package, please install it.")
+
 from lightning_habana.pytorch.plugins.io_plugin import HPUCheckpointIO
-from lightning_habana.utils.hpu_distributed import _sync_ddp_if_available
-from lightning_habana.utils.imports import _HABANA_FRAMEWORK_AVAILABLE, _TORCH_LESSER_2_3_0
+from lightning_habana.utils.hpu_distributed import _sync_hpu_processes_if_available
+from lightning_habana.utils.imports import _HABANA_FRAMEWORK_AVAILABLE
 
 if _HABANA_FRAMEWORK_AVAILABLE:
     import habana_frameworks.torch.core as htcore
@@ -57,7 +74,7 @@ if _HABANA_FRAMEWORK_AVAILABLE:
 log = logging.getLogger(__name__)
 
 
-class HPUParallelStrategy(DDPStrategy):
+class HPUParallelStrategy(ParallelStrategy):
     """Strategy for distributed training on multiple HPU devices."""
 
     strategy_name = "hpu_parallel"
@@ -66,33 +83,22 @@ class HPUParallelStrategy(DDPStrategy):
         self,
         accelerator: Optional[Accelerator] = None,
         parallel_devices: Optional[List[torch.device]] = None,
-        cluster_environment: Optional[ClusterEnvironment] = None,
         checkpoint_io: Optional[CheckpointIO] = None,
-        precision_plugin: Optional[PrecisionPlugin] = None,
-        ddp_comm_state: Optional[object] = None,
-        ddp_comm_hook: Optional[Callable] = None,
-        ddp_comm_wrapper: Optional[Callable] = None,
-        model_averaging_period: Optional[int] = None,
-        process_group_backend: Optional[str] = "hccl",
+        precision_plugin: Optional[Precision] = None,
+        cluster_environment: Optional[ClusterEnvironment] = None,
         **kwargs: Any,
     ) -> None:
-        rank_zero_warn(
-            "HPUParallelStrategy will be deprecated from lightning-habana >=1.6.0, please"
-            " make use of HPUDDPStrategy."
-        )
         super().__init__(
             accelerator=accelerator,
             parallel_devices=parallel_devices,
             cluster_environment=cluster_environment,
             checkpoint_io=checkpoint_io,
             precision_plugin=precision_plugin,
-            ddp_comm_state=ddp_comm_state,
-            ddp_comm_hook=ddp_comm_hook,
-            ddp_comm_wrapper=ddp_comm_wrapper,
-            model_averaging_period=model_averaging_period,
-            process_group_backend=process_group_backend,
-            **kwargs,
         )
+        self._process_group_backend = "hccl"
+        self._timeout = default_pg_timeout
+        self._num_nodes = 1
+        self._start_method = "spawn" if self.__class__.__name__ == "HPUParallelStrategy" else None
 
     @property
     def checkpoint_io(self) -> CheckpointIO:
@@ -108,31 +114,114 @@ class HPUParallelStrategy(DDPStrategy):
         self._checkpoint_io = io  # type: ignore
 
     def setup_environment(self) -> None:
-        if self._process_group_backend == "hccl":
-            # this env is used in overrides to check the backend initiated
-            _ws = self.cluster_environment.world_size()
-            _grank = self.cluster_environment.global_rank()
-            _lrank = self.cluster_environment.local_rank()
-            hpu_dist.initialize_distributed_hpu(world_size=_ws, rank=_grank, local_rank=_lrank)
         super().setup_environment()
+        if self.__class__.__name__ == "HPUParallelStrategy":
+            # Strategies derived from this class should handle their own distributed setups.
+            self.setup_distributed()
+        self.setup_hccl_env()
 
-    def determine_ddp_device_ids(self) -> None:
-        return None
+    def setup_hccl_env(self) -> None:
+        """Initializes the HCCL environment for distributed training on HPU devices."""
+        assert self._get_process_group_backend() == "hccl"
+        _ws = self.cluster_environment.world_size()
+        _grank = self.cluster_environment.global_rank()
+        _lrank = self.cluster_environment.local_rank()
+        hpu_dist.initialize_distributed_hpu(world_size=_ws, rank=_grank, local_rank=_lrank)
 
-    def broadcast(self, obj: object, src: int = 0) -> object:
-        if not torch.distributed.is_available():
+    def setup_distributed(self) -> None:
+        log.debug(f"{self.__class__.__name__}: setting up distributed...")
+        reset_seed()
+        self.set_world_ranks()
+        self._process_group_backend = self._get_process_group_backend()
+        assert self.cluster_environment is not None
+        _init_dist_connection(self.cluster_environment, self._process_group_backend, timeout=self._timeout)
+
+    def _get_process_group_backend(self) -> str:
+        assert self.root_device.type == "hpu"
+        return "hccl"
+
+    def set_world_ranks(self) -> None:
+        if self.cluster_environment is not None:
+            self.cluster_environment.set_global_rank(self.node_rank * self.num_processes + self.local_rank)
+            self.cluster_environment.set_world_size(self.num_nodes * self.num_processes)
+        # `LightningEnvironment.set_global_rank` will do this too, but we cannot rely on that implementation detail
+        # additionally, for some implementations, the setter is a no-op, so it's safer to access the getter
+        rank_zero_only.rank = utils_rank_zero_only.rank = self.global_rank
+
+    @property
+    def num_nodes(self) -> int:
+        return self._num_nodes
+
+    @num_nodes.setter
+    def num_nodes(self, num_nodes: int) -> None:
+        # note that world ranks is related to num_nodes, when resetting it, need to reset world ranks
+        self._num_nodes = num_nodes
+
+    @property
+    def num_processes(self) -> int:
+        return len(self.parallel_devices) if self.parallel_devices is not None else 0
+
+    @num_processes.setter
+    def num_processes(self, num_processes: int) -> None:
+        self._num_processes = num_processes
+
+    def _configure_launcher(self) -> None:
+        assert self.cluster_environment is not None
+        self._start_method = "spawn" if self._start_method is None else self._start_method
+        if self._start_method == "popen":
+            self._launcher = _SubprocessScriptLauncher(self.cluster_environment, self.num_processes, self.num_nodes)
+        else:
+            self._launcher = _MultiProcessingLauncher(self, start_method=self._start_method)
+
+    @property
+    def root_device(self) -> torch.device:
+        assert self.parallel_devices is not None
+        return self.parallel_devices[0]
+
+    def barrier(self, name: Optional[str] = None) -> None:
+        if not _distributed_is_initialized():
+            return
+        torch.distributed.barrier()
+
+    def model_to_device(self) -> None:
+        assert self.model is not None
+        self.model.to(self.root_device)
+
+    def reduce(
+        self,
+        tensor: Union[Tensor, Any],
+        group: Optional[Any] = None,
+        reduce_op: Optional[Union[ReduceOp, str]] = "mean",
+    ) -> Union[Tensor, Any]:
+        if isinstance(tensor, Tensor):
+            if tensor.device != self.root_device:
+                tensor = tensor.to(self.root_device)
+            return _sync_hpu_processes_if_available(tensor, group, reduce_op=reduce_op)
+        return tensor
+
+    def broadcast(self, obj: TBroadcast, src: int = 0) -> TBroadcast:
+        if not _distributed_is_initialized():
             return obj
 
         obj = [obj]
-        if self.global_rank != src:
-            obj = [None]
-
-        _hpu_broadcast_object_list(obj, src, group=_group.WORLD)
+        torch.distributed.broadcast_object_list(obj, src, group=_group.WORLD)
         return obj[0]
 
-    def on_after_backward(self) -> None:
-        # Break lazy accumulation of graph after fwd+bwd
+    def training_step(self, *args: Any, **kwargs: Any) -> STEP_OUTPUT:
+        output = super().training_step(*args, **kwargs)
+        if self.__class__.__name__ == "HPUParallelStrategy":
+            return self.reduce_loss_if_parallel(output)
         htcore.mark_step()
+        return output
+
+    def reduce_loss_if_parallel(
+        self, output: Union[Tensor, dict], reduce_op: Optional[Union[ReduceOp, str]] = "mean"
+    ) -> Union[Tensor, dict]:
+        if isinstance(output, dict) and "loss" in output:
+            output["loss"] = self.reduce(output["loss"], reduce_op=reduce_op)
+        elif isinstance(output, Tensor):
+            output = self.reduce(output, reduce_op=reduce_op)
+        return output
 
     def optimizer_step(
         self,
@@ -161,15 +250,6 @@ class HPUParallelStrategy(DDPStrategy):
         htcore.mark_step()
         return super().predict_step(*args, **kwargs)
 
-    def reduce(
-        self, tensor: Tensor, group: Optional[Any] = None, reduce_op: Optional[Union[ReduceOp, str]] = "mean"
-    ) -> Tensor:
-        if isinstance(tensor, Tensor):
-            if tensor.device != self.root_device:
-                tensor = tensor.to(self.root_device)
-            return _sync_ddp_if_available(tensor, group, reduce_op=reduce_op)
-        return tensor
-
     @classmethod
     def register_strategies(cls, strategy_registry: Dict) -> None:
         strategy_registry.register(
@@ -177,88 +257,3 @@ class HPUParallelStrategy(DDPStrategy):
             cls,
             description=f"{cls.__class__.__name__}",
         )
-
-
-# The code underneath is taken from PyTorch `torch/distributed/distributed_c10d.py`
-# the distributed backend and tensor type updates for habana backend is done here before broadcast
-def _hpu_broadcast_object_list(object_list, src=0, group=None, device=None):  # type: ignore
-    from torch.distributed import Backend, _rank_not_in_group, broadcast, get_backend, get_rank
-    from torch.distributed.distributed_c10d import _object_to_tensor, _tensor_to_object
-
-    if _rank_not_in_group(group):
-        return
-
-    my_rank = get_rank()
-    # Serialize object_list elements to tensors on src rank.
-    if my_rank == src:
-        tensor_list = []
-        size_list = []
-        if _TORCH_LESSER_2_3_0:
-            tensor_list, size_list = zip(*[_object_to_tensor(obj, device) for obj in object_list])
-        else:
-            tensor_list, size_list = zip(*[_object_to_tensor(obj, device, group) for obj in object_list])
-        object_sizes_tensor = torch.cat(size_list)
-    else:
-        object_sizes_tensor = torch.empty(len(object_list), dtype=torch.long)
-
-    # Current device selection.
-    # To preserve backwards compatibility, ``device`` is default to ``None``
-    # in which case we run current logic of device selection, i.e.
-    # ``current_device`` is CUDA if backend is NCCL otherwise CPU device. In the
-    # case it is not ``None`` we move the size and object tensors to be
-    # broadcasted to this device.
-    group_backend = get_backend(group)
-    is_nccl_backend = group_backend == Backend.NCCL
-    is_hpu_backend = group_backend == "hccl"
-    if device is not None:
-        if is_nccl_backend and device.type != "cuda":
-            raise ValueError("device type must be cuda for nccl backend")
-        current_device = device
-    else:
-        current_device = torch.device("cpu")
-        if is_nccl_backend:
-            # See note about using torch.cuda.current_device() here in
-            # docstring. We cannot simply use my_rank since rank == device is
-            # not necessarily true.
-            current_device = torch.device("cuda", torch.cuda.current_device())
-    if is_nccl_backend:
-        object_sizes_tensor = object_sizes_tensor.to(current_device)
-
-    elif is_hpu_backend:
-        current_device = torch.device("hpu")
-        # Workaround: HPU doesn't not support long tensors for collectives
-        if (object_sizes_tensor.type() == "torch.LongTensor") or (object_sizes_tensor.type() == "torch.hpu.LongTensor"):
-            object_sizes_tensor = object_sizes_tensor.int()
-        else:
-            print("unhandled hpu object_sizes_tensor type :: ", object_sizes_tensor.type())
-        object_sizes_tensor = object_sizes_tensor.to(current_device)
-
-    # Broadcast object sizes
-    broadcast(object_sizes_tensor, src=src, group=group)
-
-    # Concatenate and broadcast serialized object tensors
-    if my_rank == src:
-        object_tensor = torch.cat(tensor_list)
-    else:
-        object_tensor = torch.empty(
-            torch.sum(object_sizes_tensor).int().item(),
-            dtype=torch.uint8,
-        )
-
-    if is_nccl_backend or is_hpu_backend:
-        object_tensor = object_tensor.to(current_device)
-
-    broadcast(object_tensor, src=src, group=group)
-    # Deserialize objects using their stored sizes.
-    offset = 0
-    if my_rank != src:
-        for i, obj_size in enumerate(object_sizes_tensor):
-            obj_view = object_tensor[offset : offset + obj_size]
-            obj_view = obj_view.type(torch.uint8)
-            if obj_view.device != torch.device("cpu"):
-                obj_view = obj_view.cpu()
-            offset += obj_size
-            if _TORCH_LESSER_2_3_0:
-                object_list[i] = _tensor_to_object(obj_view, obj_size)
-            else:
-                object_list[i] = _tensor_to_object(obj_view, obj_size, group)